fix: many improvements. Seems stable

+++ libraries/espDMX_RDM/espDMX_RDM.h
@@ -88,7 +88,7 @@ struct dmx_ {
     uint16_t txChan;
     uint16_t txSize;
 
-    long full_uni_time;
+    uint32_t full_uni_time;
     long last_dmx_time;
     long led_timer;
     bool newDMX = false;
